Nuke `OrderedTupleConstructorStub`
